Move Provider to separate module
